Move UpdateCalendar request body parser to offer namespace
