+++ src/Http/Offer/UpdateCalendarBackwardCompatibilityRequestBodyParser.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Http\Event;
+namespace CultuurNet\UDB3\Http\Offer;
 
 use CultuurNet\UDB3\Http\Request\Body\RequestBodyParser;
 use Psr\Http\Message\ServerRequestInterface;
